refactor(iris): extract ContextStore persistence layer
Move the load/save/migrate persistence logic out of ContextStore into a
dedicated ContextPersistence class. Move the StoredState interface to a
small shared types file (contextStateTypes.ts) so persistence and
ContextStore can share the type without one importing the other's class.

Behavior is preserved verbatim, including the asymmetry covered by the
Task 0 migration safety-net test:
- load() of a current-version record strips sessions and activeSessionId
- migrate() of an older-version record preserves both via ?? fallbacks

ContextStore: 498 → 435 LOC. New files: contextPersistence.ts (62 LOC),
contextStateTypes.ts (12 LOC).
